# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UptimePulse API
  version: 0.0.1
  description: |
    API REST de UptimePulse (monitorización de uptime). Se sirve en `/docs`
    (Swagger UI) desde la propia API.

    **Autenticación.** Dos formas, ambas por `Authorization: Bearer …`:
    - un *access token* JWT de 5 min (de `/auth/register`, `/auth/login` o
      `/auth/refresh`; el refresh token viaja en una cookie httpOnly y caduca
      a los `SESSION_IDLE_TIMEOUT_MINUTES` de inactividad, 15 por defecto,
      renovándose en cada refresh);
    - una **API key** `up_…` creada por un admin en
      `/organizations/{id}/api-keys`. Actúa en nombre de la organización con
      el rol equivalente a sus scopes (`read` → readonly, `write` → editor) y
      nunca puede administrar la organización (`/me`, equipo, claves).

    **Organización activa.** Las rutas de datos actúan sobre la organización
    indicada en `X-Organization-Id`; sin cabecera, sobre la personal del
    usuario. Con API key la fija la clave y la cabecera se ignora.

    **Errores.** Siempre `{ "error": "<mensaje en español>" }`; los 400 de
    validación pueden llevar el detalle de zod en `error`. Cada respuesta
    lleva `X-Request-Id`.

    **Rate limits.** 300 peticiones/min por IP o API key en toda la API (429);
    login y registro 10/min; `/public/status`, `/invitations/{token}`
    y `/auth/username-available` 60/min.
servers:
  - url: /api
    description: Self-hosted (detrás del nginx del frontend, mismo origen)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: JWT de acceso o API key `up_…`.
  parameters:
    organizationHeader:
      name: X-Organization-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Organización activa. Sin ella, la personal del usuario.
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          description: Mensaje legible, o el detalle de validación de zod.
          oneOf: [{ type: string }, { type: object }]
...
    Role:
      type: string
      enum: [admin, editor, readonly]
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { $ref: "#/components/schemas/OrganizationSlug" }
        role: { $ref: "#/components/schemas/Role" }
    OrganizationSlug:
      type: string
      nullable: true
      pattern: "^[a-z0-9-]+$"
      maxLength: 60
      description: |
        Identificador público de la organización: hace que sus status pages
        sean alcanzables en `/status/team/<slug>/<page-slug>`. Único entre
        todas las organizaciones. null = sin URL pública propia.
    OrganizationDetail:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { $ref: "#/components/schemas/OrganizationSlug" }
        role: { $ref: "#/components/schemas/Role" }
        usage:
          type: object
          properties:
            monitors: { type: integer }
    Member:
      type: object
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        fullName: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        joinedAt: { type: string, format: date-time }
    Invitation:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
  responses:
...
    Forbidden:
      description: Sin rol suficiente, organización ajena, o acción de persona intentada con API key.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    BadRequest:
      description: Validación.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
...
  /organizations:
    get:
      tags: [Organizaciones]
      summary: Mis organizaciones y mi rol en cada una
      security: [{ bearer: [] }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Organization" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /organizations/{id}:
    get:
      tags: [Organizaciones]
      summary: Detalle con mi rol y nº de monitores
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Detalle.", content: { application/json: { schema: { $ref: "#/components/schemas/OrganizationDetail" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [Organizaciones]
      summary: Renombrar la organización o cambiar su identificador público
      description: Requiere rol admin.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                slug: { $ref: "#/components/schemas/OrganizationSlug" }
      responses:
        "200": { description: "Actualizada.", content: { application/json: { schema: { $ref: "#/components/schemas/Organization" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: "Ese identificador público ya está en uso." }
  /organizations/{id}/members:
    get:
      tags: [Organizaciones]
      summary: Miembros
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Member" } } } } }
  /organizations/{id}/members/{userId}:
    patch:
      tags: [Organizaciones]
      summary: Cambiar el rol de un miembro (admin)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [role], properties: { role: { $ref: "#/components/schemas/Role" } } } } }
      responses:
        "200": { description: "Actualizado." }
        "409": { description: "Sería degradar al último admin." }
    delete:
      tags: [Organizaciones]
      summary: Expulsar a un miembro (admin)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: "Expulsado." }
        "409": { description: "Es el último admin." }
  /organizations/{id}/invitations:
    get:
      tags: [Organizaciones]
      summary: Invitaciones pendientes (admin)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Invitation" } } } } }
    post:
      tags: [Organizaciones]
      summary: Invitar por email (admin); envía un enlace válido 7 días
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201": { description: "Invitación creada y email enviado.", content: { application/json: { schema: { $ref: "#/components/schemas/Invitation" } } } }
        "409": { description: "Ya es miembro o ya está invitado." }
        "502": { description: "No se pudo enviar el email (la invitación se descarta)." }
  /organizations/{id}/invitations/{invitationId}:
    delete:
      tags: [Organizaciones]
      summary: Anular una invitación (admin)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: "Anulada." }
  /invitations/{token}:
    get:
      tags: [Organizaciones]
      summary: Datos de una invitación (público, para la pantalla de aceptación)
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Invitación válida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizationName: { type: string }
                  email: { type: string }
                  role: { $ref: "#/components/schemas/Role" }
                  expiresAt: { type: string, format: date-time }
        "404": { description: "Inexistente, usada o caducada." }
  /invitations/{token}/accept:
    post:
      tags: [Organizaciones]
      summary: Aceptar (el email de la sesión debe coincidir con el invitado)
      security: [{ bearer: [] }]
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "Ahora eres miembro." }
        "403": { description: "La sesión no es del email invitado." }
        "404": { description: "Inexistente, usada o caducada." }
...